requirements.txt
opencv-python 
numpy
pyrealsense2-macosx
pygame
keyboard
numpy
pyautogui